Transformations: Split tests and move into `transformations.inline`
